backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Limiter
SQLAlchemy-Utils
flask-cors
psycopg2-binary
pyjwt
bcrypt
cachetools
requests
gunicorn